Silence unused-result warnings
The ignored write error in bytebuffer_flush() could result in
strange behavior that's hard to track down. The buffer is cleared
immediately after, so any screen updates would be lost but assumed
to be made successfully. This should probably return an error code
instead and *not clear the buffer*.

Ignoring the winch_fds related errors is probably not a big deal but
would be nice to log. Logging warnings in a TUI typically requires
reopening stderr to a file somewhere, so punting on that for now.