Fix loading undefined terminfo capabilities
Not all terminfo definitions include values for the strings we'd
like to load. When a terminfo string isn't defined, its offset is
set to -1 (0xFFFF). Previously we'd happily add this to the
beginning of the strings table, causing the string to be read from
one byte before the strings table.

This surfaced while testing under kitty [1], which does not define
the blink capability. The output demo was showing garbled results
for that area and it became apparent that the blink string loaded
via the xterm-kitty terminfo file was incorrect and nonsensical.

[1]: https://sw.kovidgoyal.net/kitty/