Add TB_BLINK fg attribute
Previously you set the TB_BOLD bit on the bg attributes to get
blinking text. That has been retained for compatibility.

Note that TB_UNDERLINE was changed from 0x0200 to 0x0800 and
TB_REVERSE was changed from 0x0400 to 0x2000. Both could be returned
to their prior values if binary compatibility with original termbox
is desired.