Axe tb_term global
Having a global terminal struct is the curses model but I don't
think it adds very much in terms of convenience. Further, it creates
a lot of complexity in the interface because you need convenience
versions of functions that work on the global terminal instance and
then individual instance versions. Let's see if we can avoid the
term global entirely at this layer.