tbsgr.c library for generating SGR sequences
Simplifies and expands on the original termbox model. Each SGR
attribute gets a bit in a 16bit int. Those attributes can then be
combined with a 16bit color value.

This first commit just adds attribute flags and a basic encoding
function that turns an int32 into an array of integer SGR codes.