Major tbsgr.c rework to struct instead of int val
Instead of trying to shoehorn everything into a single color value,
add a 64-bit struct with 32-bits for attribute flags and two 16-bit
integers for colors. This will let us add more flags and support up
to 65K-color mode.

The downside is that creating SGR values is more verbose and also
maybe a little more error prone. You have to remember to set the
BG/FG attribute flags and also remember the positions of the fg and
bg colors when creating the struct; or, be verbose and declare the
struct with member names.