Support retrieving terminfo caps by name
This adds a linear search of capability names to find indexes
which could be improved but moving to strings is necessary since
there's no way to load extended capabilities without referring to
them by index.

The index-based functions remain as ti_getflagi(), ti_getnumi(), and
ti_getstri() and may be used when loading legacy format
capabilities.