Fix bad extended name count calc with NULL ext strs
You can't calculate the number of extended names in the string table
by subtracting the number of extended string offsets from the total
entries in the string table. That's because extended string
capability offsets can be -1, meaning *there is no string for this
capability in the string table*.

Instead, calculate the number of extended name offsets by adding the
number of extended bools, nums, and strings.